# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SiteForge AI API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /discovery/search:
    post:
      operationId: searchBusinesses
      tags: [discovery]
      summary: Search for businesses
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessSearchInput"
      responses:
        "200":
          description: List of businesses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BusinessResult"
  /discovery/autoscan:
    post:
      operationId: autoScanArea
      tags: [discovery]
      summary: Auto-scan all businesses in an area
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AutoScanInput"
      responses:
        "200":
          description: List of businesses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BusinessResult"

  # ── AutoScan Scheduler ────────────────────────────────────────────────────
...
components:
  schemas:
...
    BusinessSearchInput:
      type: object
      required: [query, city]
      properties:
        query:
          type: string
        city:
          type: string
        country:
          type: string
        hasWebsite:
          type: ["boolean", "null"]
    AutoScanInput:
      type: object
      required: [city]
      properties:
        city:
          type: string
        country:
          type: string
        categories:
          type: array
          items:
            type: string
    BusinessResult:
      type: object
      required: [businessName, businessType, city, hasWebsite, leadScore]
      properties:
        businessName:
          type: string
        businessType:
          type: string
        address:
          type: ["string", "null"]
        city:
          type: string
        country:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        hasWebsite:
          type: boolean
        leadScore:
          type: integer
        placeId:
          type: ["string", "null"]
        googleMapsUrl:
          type: ["string", "null"]
        instagramUrl:
          type: ["string", "null"]
        facebookUrl:
          type: ["string", "null"]
        whatsappNumber:
          type: ["string", "null"]
        whatsappUrl:
          type: ["string", "null"]
        whatsappMessage:
          type: ["string", "null"]
        lat:
          type: ["number", "null"]
        lon:
          type: ["number", "null"]
...