# lib/api-zod/src/generated/types/businessResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SiteForge AI API
 * OpenAPI spec version: 0.1.0
 */

export interface BusinessResult {
  businessName: string;
  businessType: string;
  /** @nullable */
  address?: string | null;
  city: string;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  website?: string | null;
  hasWebsite: boolean;
  leadScore: number;
  /** @nullable */
  placeId?: string | null;
  /** @nullable */
  googleMapsUrl?: string | null;
  /** @nullable */
  instagramUrl?: string | null;
  /** @nullable */
  facebookUrl?: string | null;
  /** @nullable */
  whatsappNumber?: string | null;
  /** @nullable */
  whatsappUrl?: string | null;
  /** @nullable */
  whatsappMessage?: string | null;
  /** @nullable */
  lat?: number | null;
  /** @nullable */
  lon?: number | null;
}